tests/wpt/web-platform-tests/app-history/navigate/navigate-transitionWhile.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async () => {
  let promise_value = { abc: 'def' };
  let promise = Promise.resolve(promise_value);
  appHistory.onnavigate = e => e.transitionWhile(promise);
  let returned_promise = appHistory.navigate("#1");
  assert_not_equals(promise, returned_promise);
  let result = await returned_promise;
  assert_equals(result, undefined);
}, "navigate() and transitionWhile() with a fulfilled promise");
</script>
